requirements.txt
Flask==1.1.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-SSLify==0.1.5
telebot==0.0.3
certifi==2019.11.28
chardet==3.0.4
dnspython==1.16.0
gunicorn==19.10.0
idna==2.9
psycopg2==2.8.5
protobuf==3.6.1
PySocks==1.7.1
pyTelegramBotAPI==3.6.7
python-dotenv==0.11.0
requests==2.10.0
six==1.14.0
urllib3==1.25.8
pytest==5.4.2